Add enemies to board
Add a bunch of enemies to the board. The count will depend on how long
you've been playing the game.

This currently kind of works in that it adds enemies to the board at
random coordinates and adds more as time goes on (too many actually,
I'll need to reduce the count or slow down the spawn rate).

The trouble is that it looks like it's blocking on the `system` call to
`python`. This is a problem because it freezes the board and prevents
any player movement while the Python random command is working. Let's
try using Vim's built in `:python` to do this to see if that makes any
difference.

game.vim:
  Add a `ticks` script variable that tracks clock ticks and increments
  each time our game loop executes. This will allow us to keep track of
  time in a consistent way and make it easier for us to pause time when
  we pause the game. Also, I envision it being easier to deal with
  smaller numbers starting from 0 instead of a whole Unix timestamp. Got
  the idea for ticks from Tetris.vim
  (http://www.vim.org/scripts/script.php?script_id=172)

enemy.vim:
  * Set player character to a spade (like in playing cards) because when
    it's small it could kind of look like a popcorn kernel
  * Set a min and max amount of enemies that can be on the board at any
    given time because we don't want a sparse board and we also don't
    want to overwhelm the player when they've been playing for a long
    time and the spawn count increases to an insane amount.
  * Create a function that will drop an enemy at random coordinates on
    the board not on the player's coordinates.

      TODO: we should extend this function to ensure enemies can't spawn
      within a certain radius of the player so that we don't get random
      instances of being a dick.

      TODO: also ensure that we don't place an enemy on top of another
      enemy or on top of a missile.
  * Every 5 seconds update the enemies on the board and regenerate them
    such that the enemy count is withing the min and max enemy counts.