game.vim: Create `ClearBoardCell` function
Sets a board cell at a given set of coordinates to ' ' (space).

This allows us to call this new function from `ClearPlayerCell` instead
of having to pass in the empty cell string.

We need this function to be able to clear an enemy from the board when
it gets hit by a missile.