mappings#Listen: Show "Game Over" screen on `q`
When pressing `q`, in addition to quitting the game also display the
"Game Over" screen.

Without this, when you quit the game the board would stay the same as it
was just before pressing `q`. It seemed kind of confusing not to have
any feedback that you quit the game so I figured I would do this to at
least let you know that you successfully quit the game.