life#RenderLives: Add space between `|` and diamonds
Add a space after the `|` pipe character we added in
c9934c4b1192f50ae0d6930381622c58fbe7048a to put some breathing room
between it and the diamond characters.

Went back and forth on this and neither looked quite right but I think
I'm going to go with this one.