Add space padding to Lives and Score lines
Pad these 2 status lines with spaces to the right of their string values
so that their background colours extend to the full width of the game
board.

First tried `len()`, then `strlen()` when counting characters in
`padding#RightPadding()` but turns out those deal with bytes and the
player character is a multi-byte character so we would come up sort on
spaces for padding in the Lives line because the player diamond counts
for double. To get around this, used `strchars()` which gives us a count
of the characters instead of the bytes.