colors#Initialize: Set background colours
Set the background colour for the game board to be my normal colour
scheme's background colour so that people with other colour schemes can
see the game with the full set of colours I've been working with.

Set the foreground colour to white to allow it to be readable over the
dark dark grey background (otherwise we wouldn't be able to read the
"Game Over" text).