Save high scores
When you get a game over, save your high score to a file so that we can
display it on the leaderboard. The saved scores get sorted from highest
to lowest.

If we don't have your name yet (saved to a separate file), we prompt you
for it.

This means that you can only have one player (name) per installation.
Not the best scenario because maybe you have multiple people playing on
your computer, but I made a judgement call to do it this way so we don't
have to keep prompting the player for their name every time they get a
Game Over. That could get annoying fast.