leaderboard#RenderLeaderboard: Show headline and scores
Output our "Leaderboard" header. Two lines needed to be surrounded in
double quotes since they contain single quotes and backslashes needed to
be escaped in those strings.

Output player name and high score value. Not very pretty unfortunately.
Using hard-coded left padding where it would be nice to calculate
padding needed to center the name and score.

Also would like to right-align score numerals but leaving it as an
enhancement as there are more pressing things to do for now.